+++ src/app/main/channels/mobile-channels/mobile-channels.component.ts
@@ -8,6 +8,7 @@ import { Router } from '@angular/router';
 import { ChatService } from '../../../service/chat.service';
 import { NewMessageComponent } from '../../new-message/new-message.component';
 import { LoginService } from '../../../service/login.service';
+import { DirectMessageService } from '../../../service/direct-message.service';
 
 
 @Component({
@@ -20,7 +21,7 @@ import { LoginService } from '../../../service/login.service';
 export class MobileChannelsComponent  implements OnInit {
   private dialog = inject(MatDialog);
 
-  constructor(public mainService: MainServiceService, private loginService: LoginService, private router: Router, public chatService: ChatService,) {}
+  constructor(public mainService: MainServiceService, private loginService: LoginService, private router: Router, public chatService: ChatService, public directMessageService: DirectMessageService) {}
   channelListOpen: boolean = true;
   userListOpen: boolean = true;
   currentUser: any;
@@ -40,16 +41,13 @@ export class MobileChannelsComponent  implements OnInit {
     this.router.navigate(['/direct-chat', userId]); 
   }
 
-
   openDialogAddChannel() {
     this.dialog.open(AddChannelComponent);
   }
 
-
   openDialogNewMessage() {
     this.dialog.open(NewMessageComponent);
   }
-
 
   openChannels() {
     this.channelListOpen = !this.channelListOpen;
