+++ src/app/service/chat.service.ts
@@ -10,6 +10,7 @@ import { MentionUser } from '../../assets/models/mention-user.class';
 import { DialogUserChatComponent } from '../main/dialog/dialog-user-chat/dialog-user-chat.component';
 import { User } from '../../assets/models/user.class';
 import { Router } from '@angular/router';
+import { DirectMessageService } from './direct-message.service';
 @Injectable({
   providedIn: 'root',
 })
@@ -31,11 +32,6 @@ export class ChatService {
   idOfChannel: string = '';
   indexOfChannelMessage: number = 0;
   clickedUser: User = new User();
-  directMessageId: string = '';
-  newDataDirectMessage: Channel = new Channel();
-  dataDirectMessage: Channel = new Channel();
-  directMessageDocId: string = '';
-  directMessageIdIsAvailable = false;
 
   constructor(public mainService: MainServiceService, private router: Router) { }
 
@@ -209,157 +205,4 @@ export class ChatService {
     }
   }
 
-  /**
-   * Opens a user profile and initializes a chat dialog.
-   * @async
-   * @param {string} userId - The ID of the user whose profile is to be opened.
-   */
-  async openProfil(userId: string) {
-    this.clickedUser.id = userId;
-    await this.loadDirectChatUser(userId);
-    this.dialogInstance = this.dialog.open(DialogUserChatComponent);
-  }
-
-  /**
-   * Loads the direct chat user's data from Firestore.
-   * @async
-   * @param {string} userId - The ID of the user to load.
-   */
-  async loadDirectChatUser(userId: string) {
-    try {
-      const userDocRef = doc(this.firestore, 'users', userId);
-      const userDocSnap = await getDoc(userDocRef);
-      if (userDocSnap.exists()) {
-        this.clickedUser = new User(userDocSnap.data());
-        this.clickedUser.id = userId;
-      }
-    } catch (error) {
-      console.error('Fehler beim Laden der Benutzerdaten:', error);
-    }
-  }
-
-  /**
-   * Opens a direct message with a specific user by loading user data, checking message availability,
-   * pushing the direct message document to Firebase, and navigating to the message.
-   * @async
-   * @param {string} userId - The ID of the user with whom to open a direct message.
-   */
-  async openDirectMessage(userId: string) {
-    this.clickedUser.id = userId;
-    await this.loadDirectChatUser(userId);
-    await this.directMessageIsAvailable();
-    await this.pushDirectMessageDocToFirebase();
-    this.directMessageDocId = this.mainService.docId;
-    this.navigateDirectMessage(this.directMessageId);
-  }
-
-  /**
-   * Checks for the availability of a direct message by comparing messages from the clicked user and the logged-in user.
-   * If a common message is found, it sets the direct message ID and its availability status.
-   */
-  directMessageIsAvailable() {
-    this.directMessageIdIsAvailable = false;
-    this.directMessageId = '';
-    this.clickedUser.message.forEach((clickedUserMessage) => {
-      this.mainService.loggedInUser.message.forEach((loggedInUserMessage) => {
-        if (clickedUserMessage === loggedInUserMessage) {
-          let loggedInUserMessageString: string = loggedInUserMessage.toString();
-          this.directMessageId = loggedInUserMessageString;
-          this.directMessageIdIsAvailable = true;
-        }
-      });
-    });
-  }
-
-  /**
-   * Pushes a new direct message document to Firebase if no existing direct message ID is available.
-   * A new document is added and the message ID is then pushed to the user.
-   * @async
-   */
-  async pushDirectMessageDocToFirebase() {
-    if (!this.directMessageIdIsAvailable) {
-      this.newDataDirectMessage.channelUsers = [];
-      await this.mainService.addNewDocOnFirebase(
-        'direct-message',
-        this.newDataDirectMessage
-      );
-      this.pushDirectMessageIdToUser();
-    }
-  }
-
-  /**
-   * Updates the message arrays for both the logged-in and clicked users with a new direct message ID.
-   * Also updates the new direct message content in Firebase by adding both users to the message's channel users list.
-   */
-  pushDirectMessageIdToUser() {
-    this.mainService.loggedInUser.message.push(this.mainService.docId);
-    this.clickedUser.message.push(this.mainService.docId);
-    this.directMessageId = this.mainService.docId;
-    this.newDataDirectMessage.channelUsers.push(
-      new User(this.mainService.loggedInUser)
-    );
-    this.newDataDirectMessage.channelUsers.push(new User(this.clickedUser));
-    this.pushNewDirectmessageContenToFb();
-  }
-
-  /**
-   * Pushes new direct message content to Firebase by updating user documents for both the logged-in user and the clicked user,
-   * and adding a new channel document for the direct message.
-   * @async
-   */
-  async pushNewDirectmessageContenToFb() {
-    await this.mainService.addDoc(
-      'users', this.mainService.loggedInUser.id, new User(this.mainService.loggedInUser));
-    if (this.mainService.loggedInUser.id !== this.clickedUser.id) {
-      await this.mainService.addDoc('users', this.clickedUser.id, new User(this.clickedUser)
-      );
-    }
-    await this.mainService.addDoc('direct-message', this.directMessageId, new Channel(this.newDataDirectMessage)
-    );
-  }
-
-  /**
-   * Navigates to the direct chat page using the specified direct message ID.
-   * @param {string} id - The ID of the direct message to navigate to.
-   */
-  navigateDirectMessage(id: string) {
-    this.router.navigate(['/direct-chat', id]);
-  }
-
-  /**
-   * Sends a message within a direct message channel. It first loads the direct chat content, sets the new message details, and pushes the updated channel data to Firebase.
-   * @async
-   * @param {string} channelId - The ID of the direct message channel.
-   * @param {string} textContent - The text content of the message to be sent.
-   */
-  async sendMessageFromDirectMessage(channelId: string, textContent: string) {
-    await this.loadDirectChatContent(this.directMessageId);
-    this.messageChannel.message = textContent;
-    this.messageChannel.date = Date.now();
-    this.messageChannel.userId = this.mainService.loggedInUser.id;
-    this.messageChannel.userName = this.mainService.loggedInUser.name;
-    this.messageChannel.userEmail = this.mainService.loggedInUser.email;
-    this.messageChannel.userAvatar = this.mainService.loggedInUser.avatar;
-    this.dataDirectMessage.messageChannel.push(this.messageChannel);
-    await this.mainService.addDoc('direct-message', this.directMessageId, new Channel(this.dataDirectMessage)
-    );
-  }
-
-  /**
-   * Loads the content of a direct chat for a given user ID from Firestore. If the document exists, it initializes the data for direct messaging.
-   * @async
-   * @param {string} userId - The ID of the user whose direct message content is to be loaded.
-   */
-  async loadDirectChatContent(userId: string) {
-    try {
-      const userDocRef = doc(this.firestore, 'direct-message', userId);
-      const userDocSnap = await getDoc(userDocRef);
-      if (userDocSnap.exists()) {
-        this.dataDirectMessage = new Channel(userDocSnap.data());
-        this.dataDirectMessage.id = userId;
-      }
-    } catch (error) {
-      console.error('Fehler beim Laden der Benutzerdaten:', error);
-    }
-  }
 }
